change type long

+++ Fibonacci.java
@@ -9,10 +9,10 @@ public class Fibonacci{
 			System.out.println("No Argument.\nPlease enter the Input Value");
 			return ;
 		}
-        int fib1 = 0;
-        int fib2 = 1;
+        long fib1 = 0;
+        long fib2 = 1;
         for(int i = 0; i <= (start -2); i++) {
-            int fib3 = fib1 + fib2;
+            long fib3 = fib1 + fib2;
             System.out.println("F"+ i + " = " + fib1 + ", F" + (i+1) + " = " + fib2 + ", F" + (i+2) + " =  " + fib3);
             fib1 = fib2;
             fib2 = fib3;
